+++ plugins/HardwareDevices/powergraph.c
@@ -376,7 +376,7 @@ PPH_STRING RaplGraphSingleLabelYFunction(
     _In_ FLOAT Parameter
     )
 {
-    DOUBLE value = (DOUBLE)Value * (DOUBLE)Parameter;
+    FLOAT value = (FLOAT)Value * (FLOAT)Parameter;
 
     if (value != 0)
     {
